+++ glide-core/src/request_type.rs
@@ -193,6 +193,7 @@ pub enum RequestType {
     Watch = 183,
     UnWatch = 184,
     GeoSearchStore = 185,
+    SUnion = 186,
 }
 
 fn get_two_word_command(first: &str, second: &str) -> Cmd {
@@ -386,6 +387,7 @@ impl From<::protobuf::EnumOrUnknown<ProtobufRequestType>> for RequestType {
             ProtobufRequestType::LPos => RequestType::LPos,
             ProtobufRequestType::LCS => RequestType::LCS,
             ProtobufRequestType::GeoSearch => RequestType::GeoSearch,
+            ProtobufRequestType::SUnion => RequestType::SUnion,
             ProtobufRequestType::Watch => RequestType::Watch,
             ProtobufRequestType::UnWatch => RequestType::UnWatch,
             ProtobufRequestType::GeoSearchStore => RequestType::GeoSearchStore,
@@ -578,6 +580,7 @@ impl RequestType {
             RequestType::LPos => Some(cmd("LPOS")),
             RequestType::LCS => Some(cmd("LCS")),
             RequestType::GeoSearch => Some(cmd("GEOSEARCH")),
+            RequestType::SUnion => Some(cmd("SUNION")),
             RequestType::Watch => Some(cmd("WATCH")),
             RequestType::UnWatch => Some(cmd("UNWATCH")),
             RequestType::GeoSearchStore => Some(cmd("GEOSEARCHSTORE")),